CompositeImplicitAutograd compliance testing
Related to #61669.

Functions registered as CompositeImplicitAutograd MUST work for most, if not
all, backends. This means that we impose a set of constraints that make it more
difficult to write a CompositeImplicitAutograd function than writing regular
PyTorch code.

Concretely, this PR adds tests for all OpInfos that checks for
compliance. The things that get tested in this PR are:
- the op does not change the metadata of a Tensor without performing
dispatches
- the op does not call `set_` or `resize_`
- the op does not directly access the data ptr. This check is not robust
in that it will not catch all data_ptr accesses but should catch more.

The mechanism for the test is to create a new `__torch_dispatch__`
object, CompositeCompliantTensor. For each operator, we wrap all inputs
in CompositeCompliantTensor, turn on `enable_factory_dispatch` for it,
and send it through the operator.

Non-CompositeImplicitAutograd operators will pass the test because they
perform a dispatch to backend code. Here's how CompositeCompliantTensor
catches problems:
- If it sees `set_` or `resize_` getting called, it will directly error
out
- After each operation, CompositeCompliantTensor checks to make sure
that its metadata is consistent with that of the thing it is wrapping.
If the CompositeImplicitAutograd op modifes the metadata directly
(through e.g. the TensorImpl API) then the metadata will go out of sync.
- CompositeCompliantTensor's storage is a storage of same device as the
thing it is wrapping that has garbage values. These garbage values will
cause most accesses to the data_ptr to error out in some way or another.

CompositeCompliantTensor is written in an interesting way. First off,
if a view operation occurs (e.g. B = A.view_op(...)), then B.storage()
must alias A.storage() where B.storage() is CompositeCompliantTensor's
storage, NOT the storage of the tensor it is wrapping. This is an
invariant in autograd, see
https://github.com/pytorch/pytorch/issues/62182 for details. To handle
this we replay the view on A's storage and set it as B's storage.

Secondly, there are cases where the metadata is allowed to go out of
sync. I believe this is only possible with in-place view functions, like
`transpose_, t_, squeeze_, unsqueeze_`. Those are special cased.

Finally, I added a new section to `aten/src/ATen/native/README.md` about
what it means to be CompositeImplicitAutograd Compliant

Test Plan:
- run tests

ghstack-source-id: 466d36c1333d92a0019cf054948273401ad4f842
Pull Request resolved: https://github.com/pytorch/pytorch/pull/62772